fix type error

+++ librz/hash/algorithms/murmur3/murmur3.h
@@ -14,6 +14,6 @@
 
 void MurmurHash3_x86_32(const void *key, int len, size_t seed, void *out);
 void MurmurHash3_x86_128(const void *key, int len, size_t seed, void *out);
-void MurmurHash3_x64_128(const void *key, int len, size_t seed, void *out);
+void MurmurHash3_x64_128(const void *key, int len, uint64_t seed, void *out);
 
 #endif // _MURMURHASH3_H_
